Lab9_TensorFlow/requirements.txt
Flask==3.1.0
gpiozero==2.0.1
numpy==1.26.4
opencv-python==4.11.0.86
pigpio==1.78
pillow==11.2.1
requests==2.32.3
tflite-runtime==2.14.0
